refactor: reorganise project structure and rename files to match new layout
- Moved controller into new folder
- Renamed files for clarity and consistency
- Added .gitkeep files to preserve empty folder structure
