+++ pom.xml
@@ -58,7 +58,7 @@
             <!-- Default configuration for running with: mvn clean javafx:run -->
             <id>default-cli</id>
             <configuration>
-              <mainClass>com.example.cab302_project_pomodora/com.example.cab302_project_pomodora.HelloApplication</mainClass>
+              <mainClass>com.example.cab302_project_pomodora/com.qut.cab302_project_pomodora.Main</mainClass>
               <launcher>app</launcher>
               <jlinkZipName>app</jlinkZipName>
               <jlinkImageName>app</jlinkImageName>
