+++ src/main/resources/com/qut/cab302_project_pomodora/fxml/home.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.cab302_project_pomodora.HelloController">
+      fx:controller="com.qut.cab302_project_pomodora.controller.HomeController">
   <padding>
     <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
   </padding>
